fix: keep language when reload
