+++ nginx/src/script/document.js
@@ -122,3 +122,8 @@ export function clearInput(container) {
 		inputList[i].value = "";
 	}
 }
+
+history.pushState(null, null, location.href);
+window.onpopstate = function () {
+	history.go(1);
+};
